Refactored structure into many different smaller modules.
