+++ src/Oracles/QuerySynthesisFunctionOracle.jl
@@ -1,0 +1,8 @@
+
+struct QuerySynthesisFunctionOracle <: Oracle
+    f::F where F <: Function
+end
+
+function ask_oracle(oracle::QuerySynthesisFunctionOracle, query_object)
+    return oracle.f(query_object)
+end
