+++ src/Oracles/QuerySynthesisKNNOracle.jl
