+++ src/QueryStrategies/QueryStrategies.jl
@@ -1,71 +1,58 @@
 module QueryStrategies
 
-using PyCall, Memento
+using Reexport
+using Distances
+using MLKernels
+using MLLabelUtils
+using NearestNeighbors
+using Statistics
+using LinearAlgebra
+using InteractiveUtils
+using LIBSVM
+using JuMP
+using SVDD
+using PyCall
 
 const gaussian_kde = PyNULL()
 
-function __init__()
-    copy!(gaussian_kde, pyimport_conda("scipy.stats", "scipy")[:gaussian_kde])
-end
+abstract type QueryStrategy end
+abstract type PoolQs <: QueryStrategy end
 
-include("qs_base.jl")
+function qs_score end
+
+"""
+get_query_object(qs::QueryStrategy, data::Array{T, 2}, pools::Vector{Symbol}, global_indices::Vector{Int}, history::Vector{Int})
+
+# Arguments
+- `query_data`: Subset of the full data set
+- `pools`: Labels for `query_data`
+- `global_indices`: Indices of the observations in `query_data` relative to the full data set.
+- `history`: Indices of previous queries
+"""
+function get_query_object end
+
 include("qs_utils.jl")
-include("qs_subspace.jl")
+include("qs_base.jl")
 
-include("pool/TestPQs.jl")
-include("pool/RandomPQs.jl")
-include("pool/RandomOutlierPQs.jl")
-include("pool/MinimumMarginPQs.jl")
-include("pool/ExpectedMinimumMarginPQs.jl")
-include("pool/ExpectedMaximumEntropyPQs.jl")
-include("pool/MinimumLossPQs.jl")
-include("pool/HighConfidencePQs.jl")
-include("pool/DecisionBoundaryPQs.jl")
-include("pool/HybridQuerySynthesisPQs.jl")
-include("pool/NeighborhoodBasedPQs.jl")
-include("pool/BoundaryNeighborCombinationPQs.jl")
+include("SequentialQueryStrategies/SequentialQueryStrategies.jl")
+@reexport using .SequentialQueryStrategies
 
-include("query_synthesis/query_synthesis_base.jl")
-include("query_synthesis/query_synthesis_utils.jl")
-include("query_synthesis/TestQss.jl")
-include("query_synthesis/RandomQss.jl")
-include("query_synthesis/RandomOutlierQss.jl")
-include("query_synthesis/DecisionBoundaryQss.jl")
-include("query_synthesis/NaiveExplorativeMarginQss.jl")
-include("query_synthesis/ExplorativeMarginQss.jl")
+include("QuerySynthesisStrategies/QuerySynthesisStrategies.jl")
+@reexport using .QuerySynthesisStrategies
+
+include("SubspaceQueryStrategies/SubspaceQueryStrategies.jl")
+@reexport using .SubspaceQueryStrategies
+
+
+function __init__()
+    copy!(gaussian_kde, pyimport_conda("scipy.stats", "scipy").gaussian_kde)
+end
 
 export
     QueryStrategy,
     PoolQs,
-    SubspaceQs,
-    DataBasedPQs, ModelBasedPQs, HybridPQs,
-    QuerySynthesisStrategy,
-    SubspaceQueryStrategy,
-    DataBasedQss, ModelBasedQss, HybridQss,
 
-    # pool based query strategies
-    # data-based query strategies
-    TestPQs, RandomPQs, MinimumMarginPQs, ExpectedMinimumMarginPQs, ExpectedMaximumEntropyPQs,
-    MinimumLossPQs,
-    # model-based query strategies
-    RandomOutlierPQs, HighConfidencePQs, DecisionBoundaryPQs,
-    # hybrid query strategies
-    HybridQuerySynthesisPQs, NeighborhoodBasedPQs, BoundaryNeighborCombinationPQs,
-
-    # query synthesis query query strategies
-    TestQss, RandomQss, RandomOutlierQss,
-    DecisionBoundaryQss, NaiveExplorativeMarginQss, ExplorativeMarginQss,
-
-
-    # query synthesis optimizers
-    QuerySynthesisOptimizer,
-    ParticleSwarmOptimization, EvolutionaryOptimization, BlackBoxOptimization,
-
-    get_query_object,
     qs_score,
     initialize_qs,
-    filter_array,
-    multi_kde, KDEException, MissingLabelTypeException,
-    estimate_boundary_shift_epsilon, data_limit_epsilon, data_boundaries
-
+    get_query_object
 end
