+++ src/QueryStrategies/SequentialQueryStrategies/HighConfidencePQs.jl
@@ -11,7 +11,7 @@ struct HighConfidencePQs <: ModelBasedPQs
 end
 
 function qs_score(qs::HighConfidencePQs, x::Array{T, 2}, labels::Dict{Symbol, Array{Int, 1}})::Array{Float64, 1} where T <: Real
-    return SVDD.predict(qs.occ, x)
+    return predict(qs.occ, x)
 end
 
 function qs_score(qs::HighConfidencePQs,
