+++ src/result.jl
@@ -51,7 +51,8 @@ function get_worker_info(;debug = false)
     worker_info = Dict{Symbol, String}()
     worker_info[:julia_version] = string(VERSION)
     debug && (worker_info[:installed_packages] = join( ["$k [$v]" for (k,v) in Pkg.installed()], ','))
-    cmd = Sys.iswindows() ? `cmd /c git -C $(@__DIR__) rev-parse HEAD` : `git -C $(@__DIR__) rev-parse HEAD`
+    dir = @__DIR__
+    cmd = Sys.iswindows() ? `cmd /c git -C $(dir) rev-parse HEAD` : `git -C $(@__DIR__) rev-parse HEAD`
     worker_info[:git_commit] = strip(string(read(cmd, String)))
     worker_info[:utc_time] = string(Dates.now(Dates.UTC))
     worker_info[:hostname] = gethostname()
