+++ test/QueryStrategies/query_synthesis_optimization_test.jl
@@ -1,11 +1,11 @@
-
+OPT = OneClassActiveLearning.QueryStrategies.QuerySynthesisStrategies.Optimization
 @testset "query synthesis optimization" begin
     @testset "particle swarm optimization" begin
         for d in [1, 2, 10]
             @testset "$d dim" begin
                 Random.seed!(0)
                 f(x) = vec(sum(abs.(x), dims=1))
-                x_opt, _ = OneClassActiveLearning.QueryStrategies.pso(f, -ones(d), ones(d))
+                x_opt, _ = OPT.pso(f, -ones(d), ones(d))
                 @test size(x_opt) == (d,)
                 @test all(f(reshape(x_opt, 1, d)) .< 0.015 * d)
             end
@@ -20,7 +20,7 @@
                 @testset "$(typeof(opt)): $d dim" begin
                     Random.seed!(0)
                     f(x) = -vec(sum(abs.(x), dims=1))
-                    x_opt = OneClassActiveLearning.QueryStrategies.query_synthesis_optimize(f, opt, randn(d, 10), fill(:U, 10))
+                    x_opt = OPT.query_synthesis_optimize(f, opt, randn(d, 10), fill(:U, 10))
                     @test size(x_opt) == (d, 1)
                 end
             end
