+++ test/QueryStrategies/query_synthesis_utils_test.jl
@@ -1,6 +1,6 @@
 
 @testset "query synthesis utils" begin
-    UTILS = OneClassActiveLearning.QueryStrategies
+    UTILS = OneClassActiveLearning.QueryStrategies.QuerySynthesisStrategies
     @test_throws ArgumentError UTILS.check_epsilon(-1)
     @test UTILS.check_epsilon(0) === nothing
     @test_throws ArgumentError UTILS.check_epsilon([-1])
@@ -25,6 +25,6 @@
     @test all(test_minima .<= test_data .<= test_maxima)
     test_data = UTILS.rand_in_hyper_rect(test_minima, test_maxima, -0.1)
     @test all(test_minima .<= test_data .<= test_maxima)
-    test_data = UTILS.rand_in_hyper_rect(test_minima, test_maxima, [1.0; 1.0])
+    test_data = UTILS.rand_in_hyper_rect(test_minima, test_maxima, [-0.11; -0.12])
     @test all(test_minima .<= test_data .<= test_maxima)
 end
